refactor: centralize voice helpers
